tests/wpt/tests/html/canvas/element/path-objects/2d.path.roundrect.radius.noargument.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.roundrect.radius.noargument</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.roundrect.radius.noargument</h1>
<p class="desc">Check that roundRect draws a rectangle when no radii are provided.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Check that roundRect draws a rectangle when no radii are provided.");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.roundRect(10, 10, 80, 30);
  ctx.fillStyle = '#0f0';
  ctx.fill();
  // upper left corner (10, 10)
  _assertPixel(canvas, 10,9, 255,0,0,255);
  _assertPixel(canvas, 9,10, 255,0,0,255);
  _assertPixel(canvas, 10,10, 0,255,0,255);

  // upper right corner (89, 10)
  _assertPixel(canvas, 90,10, 255,0,0,255);
  _assertPixel(canvas, 89,9, 255,0,0,255);
  _assertPixel(canvas, 89,10, 0,255,0,255);

  // lower right corner (89, 39)
  _assertPixel(canvas, 89,40, 255,0,0,255);
  _assertPixel(canvas, 90,39, 255,0,0,255);
  _assertPixel(canvas, 89,39, 0,255,0,255);

  // lower left corner (10, 30)
  _assertPixel(canvas, 9,39, 255,0,0,255);
  _assertPixel(canvas, 10,40, 255,0,0,255);
  _assertPixel(canvas, 10,39, 0,255,0,255);

});
</script>
